compatibility with latest VST and AU SDKs
update MacOS architecture and C++ standard to allow M1 universal binary builds

+++ PluginTemplate/project_source/source/vst_source/vst3plugin.h
@@ -783,7 +783,7 @@ public:
     {
         if (sampleOffset == 0 && pluginCore)
         {
-            uint32_t count = proxyMIDIEvents.size();
+            uint32_t count = (uint32_t)proxyMIDIEvents.size();
             for (uint32_t i = 0; i < count; i++)
             {
                 pluginCore->processMIDIEvent(proxyMIDIEvents[i]);
